+++ emgapimetadata/management/commands/import_summary.py
@@ -5,7 +5,7 @@ import os
 import csv
 import logging
 
-from emgapi import models as emg_models
+from emgapimetadata import models as m_models
 
 from ..lib import EMGBaseCommand
 
@@ -14,6 +14,10 @@ logger = logging.getLogger(__name__)
 
 class Command(EMGBaseCommand):
 
+    def add_arguments(self, parser):
+        super(Command, self).add_arguments(parser)
+        parser.add_argument('suffix', nargs='?', type=str, default='.go_slim')
+
     def populate_from_accession(self, options):
         logger.info("Found %d" % len(self.obj_list))
         for o in self.obj_list:
@@ -21,34 +25,77 @@ class Command(EMGBaseCommand):
 
     def find_path(self, obj, options):
         rootpath = options.get('rootpath', None)
+        self.suffix = options.get('suffix', None)
+
         res = os.path.join(rootpath, obj.result_directory)
         logger.info("Scanning path: %s" % res)
         if os.path.exists(res):
             if os.path.isdir(res):
                 for root, dirs, files in os.walk(res, topdown=False):
                     for name in files:
-                        if name in ('new.summary',):
-                            with open(os.path.join(root, name)) as csvfile:
-                                reader = csv.reader(csvfile, delimiter='\t')
-                                self.import_summary(reader, obj)
+                        if name.endswith(self.suffix):
+                            _f = os.path.join(root, name)
+                            logger.info("Found: %s" % _f)
+                            with open(_f) as csvfile:
+                                reader = csv.reader(csvfile, delimiter=',')
+                                self.load_from_summary_file(
+                                    reader, obj.accession,
+                                    obj.pipeline.release_version
+                                )
+                            continue
             elif os.path.isfile(res):
                 raise NotImplementedError("Give path to directory.")
         else:
-            raise NotImplementedError("Path %r doesn't exist." % res)
+            raise NotImplementedError("Path '%r' doesn't exist." % res)
 
-    def import_summary(self, reader, job):
+    def load_from_summary_file(self, reader, accession, pipeline):  # noqa
+        if self.suffix == '.go_slim':
+            run = m_models.AnalysisJobGoSlimTerm()
+        elif self.suffix == '.go':
+            run = m_models.AnalysisJobGoTerm()
+        elif self.suffix == '.ipr':
+            run = m_models.AnalysisJobInterProTerm()
+        run.accession = accession
+        run.pipeline_version = pipeline
+        new_anns = []
         anns = []
         for row in reader:
             try:
-                var = emg_models.VariableNames.objects.get(var_name=row[0])
-            except emg_models.VariableNames.DoesNotExist:
-                var = emg_models.VariableNames(var_name=row[0])
-                var.save()
-                # becuase PK is not AutoField
-                var = emg_models.VariableNames.objects.get(var_name=row[0])
-            job_ann = emg_models.AnalysisJobAnn()
-            job_ann.job = job
-            job_ann.var = var
-            job_ann.var_val_ucv = row[1]
-            anns.append(job_ann)
-        emg_models.AnalysisJobAnn.objects.bulk_create(anns)
+                row[0].lower().startswith('go:')
+            except KeyError:
+                pass
+            else:
+                ann = None
+                try:
+                    ann = m_models.GoTerm.objects.get(accession=row[0])
+                except m_models.GoTerm.DoesNotExist:
+                    ann = m_models.GoTerm(
+                        accession=row[0],
+                        description=row[1],
+                        lineage=row[2],
+                    )
+                    new_anns.append(ann)
+                if ann is not None:
+                    anns.append(ann)
+                    if self.suffix == '.go_slim':
+                        rann = m_models.AnalysisJobGoSlimTermAnnotation(
+                            count=row[3],
+                            go_term=ann
+                        )
+                        run.go_slim.append(rann)
+                    elif self.suffix == '.go':
+                        rann = m_models.AnalysisJobGoTermAnnotation(
+                            count=row[3],
+                            go_term=ann
+                        )
+                        run.go_terms.append(rann)
+                    elif self.suffix == '.ipr':
+                        rann = m_models.AnalysisJobInterProTermAnnotation(
+                            count=row[3],
+                            interpro_term=ann
+                        )
+                        run.interpro_terms.append(rann)
+        if len(anns) > 0:
+            if len(new_anns) > 0:
+                m_models.GoTerm.objects.insert(new_anns)
+            run.save()
